Use logging instead of flags and print statements
This is a fairly direct conversion of the code to use logging. It
could be made more Pythonic.

This also gets rid of the ERROR_LOG and DEBUG_LOG variables for
redirecting error and debug messages. They could only be used when
importing smatch as a library and both pointed to stderr by
default. The logging messages are prefixed by DEBUG, ERROR, etc., so
if anything they are now easier to distinguish. Getting support for
separate streams would now require some logging Handlers and Filters,
which are a pain to set up.

Resolves #7